Add the basic CatalogApp Composable in the MainActivity to display the Catalog Content
The MainActivity is the Single Activity for this Application. It will display
all necessary screens, depending on the current selected route.
It will also behave differently depending on the current screen expension.
If the screen is not expanded (portrait mode for exemple), it will display
the catalog list as root destination, then display any selected destination
full screen.
If the screen is expanded (landscape mode for exemple), it will display
both the catalog list alongside the selected content. The Catalog list will
then act as a bottomBar or an always opened drawer.
In both cases, the selection of an item in the Catalog list will navigate to
another destination.