Add an animation when selecting/unselecting an item in the catalog list while in expanded screen
When in expanded screen, by default, the Catalog List takes up a certain portion of the screen
and the content will take the rest. With this animation, the Catalog List will take up the whole
screen if the current navigation destination is the "root" destination (ie: we are not navigating
to any existing navigation). Then, when navigating to any destination, the Catalog List width
will animate to the given portion of the screen and the content will animate as well to take
up all the remaining space.